add is last state
